requirements.txt
Flask==2.2.2
line-bot-sdk==2.0.0
google-generativeai==0.1.3